requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.7.9
charset-normalizer==3.4.2
Django==4.2.23
django-redis==6.0.0
djangorestframework==3.16.0
idna==3.10
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0